removed unused components, moved components into specif dirs
