+++ src/components/routes/ProfessorRoute.js
@@ -1,11 +1,16 @@
 import { Navigate, Outlet, useLocation } from "react-router-dom";
-import { useAuth } from "../contexts/AuthContext";
+import { useAuth } from "../../hooks/useAuth";
 
 export default function ProfessorRoute({ redirectTo = "/professor" }) {
 	const { loading, isProfessor } = useAuth();
 	const location = useLocation();
 
-	if (loading) return <div className="p-6 font-medium text-md text-center">Checking access…</div>;
+	if (loading)
+		return (
+			<div className="p-6 font-medium text-md text-center">
+				Checking access…
+			</div>
+		);
 
 	return isProfessor ? (
 		<Outlet />
